Fixed some warnings on DSUtil

+++ Projects/TMProject/CFrame.cpp
@@ -500,7 +500,7 @@ void CFrame::UpdateFrames(D3DXMATRIX* matCur)
 
 				m_pParentSkin->m_pOwner->m_AlphaColor = vecOut3.x;
 				m_pParentSkin->m_pOwner->m_nAlpha = vecOut3.x;
-				m_pParentSkin->m_pOwner->m_bAlphaObj = vecOut3.z;
+				m_pParentSkin->m_pOwner->m_bAlphaObj = static_cast<int>(vecOut3.z);
 
 				vecInFinal = { 0.40000001f, 0.2f, 0.0f };
